Simplify Hilla version
