+++ assembly/src/assembly/hilla-skeleton.xml
@@ -11,7 +11,7 @@
             <directory>${basedir}/target/generated-sources/hilla</directory>
             <excludes>
                 <exclude>com/example/application/base/**</exclude>
-                <exclude>com/example/application/taskmanagement/ui/**</exclude>
+                <exclude>com/example/application/examplefeature/ui/**</exclude>
             </excludes>
         </fileSet>
         <fileSet>
